+++ src/settings/MappingProfiles/ViewMappingProfile/ViewMappingProfile.js
@@ -62,6 +62,7 @@ import {
   marcFieldProtectionSettingsShape,
   showActionMenu,
   isMARCType,
+  BASE_URLS,
 } from '../../../utils';
 
 import sharedCss from '../../../shared.css';
@@ -71,7 +72,7 @@ import sharedCss from '../../../shared.css';
 export class ViewMappingProfile extends Component {
   static manifest = Object.freeze({
     initializedFilterConfig: { initialValue: false },
-    mappingProfile: {
+    mappingProfileView: {
       type: 'okapi',
       path: 'data-import-profiles/mappingProfiles/:{id}',
       params: { withRelations: true },
@@ -91,7 +92,7 @@ export class ViewMappingProfile extends Component {
   static propTypes = {
     stripes: stripesShape.isRequired,
     resources: PropTypes.shape({
-      mappingProfile: PropTypes.shape({
+      mappingProfileView: PropTypes.shape({
         hasLoaded: PropTypes.bool.isRequired,
         records: PropTypes.arrayOf(PropTypes.object),
       }),
@@ -133,7 +134,7 @@ export class ViewMappingProfile extends Component {
       parentResources,
     } = this.props;
 
-    const mappingProfile = resources.mappingProfile || {};
+    const mappingProfile = resources.mappingProfileView || {};
     const marcFieldProtectionFields = parentResources.marcFieldProtectionSettings.records || [];
     const [record] = mappingProfile.records || [];
 
@@ -177,7 +178,7 @@ export class ViewMappingProfile extends Component {
         entity={this}
         menu={menu}
         recordId={record?.id}
-        baseUrl="/settings/data-import/mapping-profiles"
+        baseUrl={BASE_URLS.MAPPING_PROFILE}
       />
     );
   };
@@ -284,6 +285,7 @@ export class ViewMappingProfile extends Component {
         history={history}
         location={location}
         recordId={id}
+        baseUrl={BASE_URLS.MAPPING_PROFILE}
       >
         <FullScreenView
           data-test-mapping-profile-details