+++ src/settings/MappingProfiles/detailsSections/edit/OrderDetailSection/CostDetails.js
@@ -13,12 +13,16 @@ import {
 } from '@folio/stripes/components';
 import { TypeToggle } from '@folio/stripes-acq-components';
 
-import { AcceptedValuesField } from '../../../../../components';
+import {
+  AcceptedValuesField,
+  WithValidation,
+} from '../../../../../components';
 
 import { TRANSLATION_ID_PREFIX } from '../../constants';
 import {
   getBoolFieldName,
   getFieldName,
+  renderFieldLabelWithInfo,
 } from '../../utils';
 import { BOOLEAN_ACTIONS } from '../../../../../utils';
 
@@ -28,6 +32,23 @@ export const CostDetails = ({
 }) => {
   const [isSetExchangeDisabled, setIsSetExchangeDisabled] = useState(useSetExchange === BOOLEAN_ACTIONS.ALL_FALSE);
 
+  const physicalUnitPriceLabel = renderFieldLabelWithInfo(
+    `${TRANSLATION_ID_PREFIX}.order.costDetails.field.physicalUnitPrice`,
+    `${TRANSLATION_ID_PREFIX}.order.costDetails.field.physicalUnitPrice.info`,
+  );
+  const quantityPhysicalLabel = renderFieldLabelWithInfo(
+    `${TRANSLATION_ID_PREFIX}.order.costDetails.field.quantityPhysical`,
+    `${TRANSLATION_ID_PREFIX}.order.costDetails.field.physicalUnitPrice.info`,
+  );
+  const electronicUnitPriceLabel = renderFieldLabelWithInfo(
+    `${TRANSLATION_ID_PREFIX}.order.costDetails.field.electronicUnitPrice`,
+    `${TRANSLATION_ID_PREFIX}.order.costDetails.field.electronicUnitPrice.info`,
+  );
+  const quantityElectronicLabel = renderFieldLabelWithInfo(
+    `${TRANSLATION_ID_PREFIX}.order.costDetails.field.quantityElectronic`,
+    `${TRANSLATION_ID_PREFIX}.order.costDetails.field.electronicUnitPrice.info`,
+  );
+
   return (
     <Accordion
       id="cost-details"
@@ -35,28 +56,40 @@ export const CostDetails = ({
     >
       <Row left="xs">
         <Col xs={4}>
-          <Field
-            component={TextField}
-            label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.costDetails.field.physicalUnitPrice`} />}
-            name={getFieldName(50)}
-            type="number"
-          />
+          <WithValidation>
+            {validation => (
+              <Field
+                component={TextField}
+                label={physicalUnitPriceLabel}
+                name={getFieldName(50)}
+                validate={[validation]}
+              />
+            )}
+          </WithValidation>
         </Col>
         <Col xs={4}>
-          <Field
-            component={TextField}
-            label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.costDetails.field.quantityPhysical`} />}
-            name={getFieldName(51)}
-            type="number"
-          />
+          <WithValidation>
+            {validation => (
+              <Field
+                component={TextField}
+                label={quantityPhysicalLabel}
+                name={getFieldName(51)}
+                validate={[validation]}
+              />
+            )}
+          </WithValidation>
         </Col>
         <Col xs={4}>
-          <Field
-            component={TextField}
-            label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.costDetails.field.additionalCoast`} />}
-            name={getFieldName(52)}
-            type="number"
-          />
+          <WithValidation>
+            {validation => (
+              <Field
+                component={TextField}
+                label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.costDetails.field.additionalCoast`} />}
+                name={getFieldName(52)}
+                validate={[validation]}
+              />
+            )}
+          </WithValidation>
         </Col>
       </Row>
       <Row left="xs">
@@ -69,6 +102,7 @@ export const CostDetails = ({
             optionLabel="label"
             wrapperLabel={`${TRANSLATION_ID_PREFIX}.wrapper.acceptedValues`}
             acceptedValuesList={currenciesOptions}
+            required
           />
         </Col>
         <Col xs={4}>
@@ -86,39 +120,55 @@ export const CostDetails = ({
           />
         </Col>
         <Col xs={4}>
-          <Field
-            component={TextField}
-            label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.costDetails.field.setExchangeRate`} />}
-            name={getFieldName(55)}
-            type="number"
-            disabled={isSetExchangeDisabled}
-          />
+          <WithValidation>
+            {validation => (
+              <Field
+                component={TextField}
+                label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.costDetails.field.setExchangeRate`} />}
+                name={getFieldName(55)}
+                validate={[validation]}
+                disabled={isSetExchangeDisabled}
+              />
+            )}
+          </WithValidation>
         </Col>
       </Row>
       <Row left="xs">
         <Col xs={3}>
-          <Field
-            component={TextField}
-            label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.costDetails.field.electronicUnitPrice`} />}
-            name={getFieldName(56)}
-            type="number"
-          />
+          <WithValidation>
+            {validation => (
+              <Field
+                component={TextField}
+                label={electronicUnitPriceLabel}
+                name={getFieldName(56)}
+                validate={[validation]}
+              />
+            )}
+          </WithValidation>
         </Col>
         <Col xs={3}>
-          <Field
-            component={TextField}
-            label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.costDetails.field.quantityElectronic`} />}
-            name={getFieldName(57)}
-            type="number"
-          />
+          <WithValidation>
+            {validation => (
+              <Field
+                component={TextField}
+                label={quantityElectronicLabel}
+                name={getFieldName(57)}
+                validate={[validation]}
+              />
+            )}
+          </WithValidation>
         </Col>
         <Col xs={3}>
-          <Field
-            component={TextField}
-            label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.costDetails.field.discount`} />}
-            name={getFieldName(58)}
-            type="number"
-          />
+          <WithValidation>
+            {validation => (
+              <Field
+                component={TextField}
+                label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.costDetails.field.discount`} />}
+                name={getFieldName(58)}
+                validate={[validation]}
+              />
+            )}
+          </WithValidation>
         </Col>
         <Col xs={3}>
           <Field