+++ src/settings/MappingProfiles/detailsSections/edit/OrderDetailSection/Location.js
@@ -11,7 +11,10 @@ import {
   Col,
 } from '@folio/stripes/components';
 
-import { AcceptedValuesField } from '../../../../../components';
+import {
+  AcceptedValuesField,
+  WithValidation,
+} from '../../../../../components';
 
 import {
   TRANSLATION_ID_PREFIX,
@@ -22,6 +25,7 @@ import {
   getSubfieldName,
   onAdd,
   onRemove,
+  renderFieldLabelWithInfo,
 } from '../../utils';
 
 export const Location = ({
@@ -31,6 +35,18 @@ export const Location = ({
   okapi,
 }) => {
   const locationsFieldIndex = 61;
+  const locationLabel = renderFieldLabelWithInfo(
+    `${TRANSLATION_ID_PREFIX}.order.location.field.name`,
+    `${TRANSLATION_ID_PREFIX}.order.location.field.name.info`,
+  );
+  const quantityPhysicalLabel = renderFieldLabelWithInfo(
+    `${TRANSLATION_ID_PREFIX}.order.location.field.quantityPhysical`,
+    `${TRANSLATION_ID_PREFIX}.order.location.field.quantityPhysical.info`,
+  );
+  const quantityElectronicLabel = renderFieldLabelWithInfo(
+    `${TRANSLATION_ID_PREFIX}.order.location.field.quantityElectronic`,
+    `${TRANSLATION_ID_PREFIX}.order.location.field.quantityElectronic.info`,
+  );
 
   return (
     <Accordion
@@ -48,7 +64,7 @@ export const Location = ({
               <Col xs={4}>
                 <AcceptedValuesField
                   component={TextField}
-                  label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.location.field.name`} />}
+                  label={locationLabel}
                   name={getSubfieldName(locationsFieldIndex, 0, index)}
                   optionValue="name"
                   optionLabel="name"
@@ -64,20 +80,28 @@ export const Location = ({
                 />
               </Col>
               <Col xs={4}>
-                <Field
-                  component={TextField}
-                  label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.location.field.quantityPhysical`} />}
-                  name={getSubfieldName(locationsFieldIndex, 1, index)}
-                  type="number"
-                />
+                <WithValidation>
+                  {validation => (
+                    <Field
+                      component={TextField}
+                      label={quantityPhysicalLabel}
+                      name={getSubfieldName(locationsFieldIndex, 1, index)}
+                      validate={[validation]}
+                    />
+                  )}
+                </WithValidation>
               </Col>
               <Col xs={4}>
-                <Field
-                  component={TextField}
-                  label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.order.location.field.quantityElectronic`} />}
-                  name={getSubfieldName(locationsFieldIndex, 2, index)}
-                  type="number"
-                />
+                <WithValidation>
+                  {validation => (
+                    <Field
+                      component={TextField}
+                      label={quantityElectronicLabel}
+                      name={getSubfieldName(locationsFieldIndex, 2, index)}
+                      validate={[validation]}
+                    />
+                  )}
+                </WithValidation>
               </Col>
             </Row>
           );