+++ src/utils/craftLayerURL.js
@@ -1,17 +1,21 @@
+import { LAYER_TYPES } from './constants';
+
 /**
  * Description: TBD
  *
- * @param {object} location comes from withRouter
- * @param {string} type layer type
+ * @param {object} baseUrl base location url
+ * @param {string} layerType layer type
+ * @param {string} seach search query
+ * @param {string} recordId id of the record
  * @return {string} URL created
  */
-export const createLayerURL = (location, type) => {
-  const {
-    pathname,
-    search,
-  } = location;
+export const createLayerURL = ({
+  baseUrl,
+  layerType,
+  search,
+  recordId = '',
+}) => {
+  const basicPath = `${baseUrl}/${layerType}`;
 
-  const url = `${pathname}${search}`;
-
-  return `${url}${url.includes('?') ? '&' : '?'}layer=${type}`;
+  return layerType === LAYER_TYPES.CREATE ? `${basicPath}${search}` : `${basicPath}/${recordId}${search}`;
 };